pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pro-trader"
version = "1.0.0"
description = "Pro-Trader: Plugin-based Autonomous Trading Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core (minimal for plugin framework)
    "pandas>=2.3.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "typer>=0.21.0",
    "typing-extensions>=4.14.0",
    "yfinance>=0.2.63",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Full install (everything)
all = [
    "pro-trader[agents]",
    "pro-trader[brokers]",
    "pro-trader[monitors]",
    "pro-trader[dashboard]",
]

# Multi-agent LLM analysis (LangGraph + LLM providers)
agents = [
    "langchain-core>=0.3.81",
    "langchain-anthropic>=0.3.15",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "rank-bm25>=0.2.2",
]

# Broker integrations
brokers = [
    "alpaca-trade-api>=3.0.0",
]

# Background monitors
monitors = [
    "feedparser>=6.0.0",
]

# Dashboard
dashboard = [
    "fastapi>=0.100.0",
    "uvicorn>=0.25.0",
]

# Legacy (full tradingagents compatibility)
legacy = [
    "pro-trader[agents]",
    "backtrader>=1.9.78.123",
    "chainlit>=2.5.5",
    "parsel>=1.10.0",
    "questionary>=2.1.0",
    "redis>=6.2.0",
    "setuptools>=80.9.0",
    "stockstats>=0.6.5",
    "tqdm>=4.67.1",
]

# Development
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
# New plugin-based CLI
pro-trader = "pro_trader.cli.app:main"
# Legacy CLI (keep backward compatibility)
tradingagents = "cli.main:app"

# ── Plugin Entry Points ─────────────────────────────────────────────────────
# Third-party plugins register here too:
#   [project.entry-points."pro_trader.data"]
#   polygon = "pro_trader_polygon:PolygonPlugin"

[project.entry-points."pro_trader.data"]
realtime = "pro_trader.plugins.data.realtime_plugin:RealtimePlugin"
yfinance = "pro_trader.plugins.data.yfinance_plugin:YFinancePlugin"
futures = "pro_trader.plugins.data.futures_plugin:FuturesPlugin"

[project.entry-points."pro_trader.analysts"]
flash = "pro_trader.plugins.analysts.flash_analyst:FlashAnalyst"
macro = "pro_trader.plugins.analysts.macro_analyst:MacroAnalyst"
pulse = "pro_trader.plugins.analysts.pulse_analyst:PulseAnalyst"

[project.entry-points."pro_trader.strategies"]
cooper_scorer = "pro_trader.plugins.strategies.cooper_scorer:CooperScorer"

[project.entry-points."pro_trader.brokers"]
alpaca = "pro_trader.plugins.brokers.alpaca_broker:AlpacaBrokerPlugin"

[project.entry-points."pro_trader.notifiers"]
console = "pro_trader.plugins.notifiers.console_notifier:ConsoleNotifier"
discord = "pro_trader.plugins.notifiers.discord_notifier:DiscordNotifier"

[project.entry-points."pro_trader.monitors"]
news = "pro_trader.plugins.monitors.news_monitor_plugin:NewsMonitorPlugin"
fomc = "pro_trader.plugins.monitors.fomc_monitor_plugin:FOMCMonitorPlugin"
futures_monitor = "pro_trader.plugins.monitors.futures_monitor_plugin:FuturesMonitorPlugin"

[project.entry-points."pro_trader.risk"]
circuit_breaker = "pro_trader.plugins.risk.circuit_breaker_plugin:CircuitBreakerPlugin"

[tool.setuptools.packages.find]
include = ["pro_trader*", "tradingagents*", "cli*"]
